% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_transcript_reference.R
\name{create_transcript_reference}
\alias{create_transcript_reference}
\title{Create transcriptome reference using Bambu}
\usage{
create_transcript_reference(bam_files, annotation_gtf, genome_seq, output_dir)
}
\arguments{
\item{bam_files}{Vector of paths to BAM files (aligned reads)}

\item{annotation_gtf}{Path to known GTF file (e.g., GENCODE)}

\item{genome_seq}{Path to reference genome FASTA (e.g., GRCh38.fa)}

\item{output_dir}{Directory to save outputs}
}
\value{
A SummarizedExperiment object containing the new transcript annotations
}
\description{
Create transcriptome reference using Bambu
}
