% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decompose_celltype.R
\name{decompose_cell_types_bisque}
\alias{decompose_cell_types_bisque}
\title{Cell type decomposition using BisqueRNA}
\usage{
decompose_cell_types_bisque(
  expr_matrix,
  reference_sce,
  out_dir,
  reference_colname = "cell_type"
)
}
\arguments{
\item{expr_matrix}{Matrix of bulk expression (genes x samples)}

\item{reference_sce}{SingleCellExperiment or Seurat object as reference}

\item{reference_colname}{Column in reference metadata containing cell type labels}
}
\value{
A data.frame with estimated cell type proportions (samples x cell types)
}
\description{
Cell type decomposition using BisqueRNA
}
